[PERF] stock: skip merging quants upon opening the inventory adjustment view
Originally, the inventory adjustment view will attempt to merge quants
eventhough this merging is done in the scheduler. In this PR we allow a
way to skip the function _quant_tasks which performs merging. This is
done through a system parameter.

Merging quants has been seen to cause deadlocks in cases where multiple
users are visiting the same view at the same time. Skipping this process
will make sure to only merge quants through the scheduler. This will
speedup opening the quants views and could also be helpful in debugging.

opw-4226821

closes odoo/odoo#189297

X-original-commit: 83e8e975ad0a923e5dd9f7e9c8fb2f6d9124e8c8
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Mohamed Kasem (kasm) <kasm@odoo.com>